+++ R/cre.R
@@ -158,15 +158,13 @@ cre <- function(y, z, X,
   # Estimate ITE
   if (is.null(ite)) {
     logger::log_info("Estimating ITE...")
-    ite_list_dis <- estimate_ite(y = y_dis, z = z_dis, X = X_dis,
+    ite_dis <- estimate_ite(y = y_dis, z = z_dis, X = X_dis,
                       ite_method = getElement(method_params, "ite_method_dis"),
                       include_ps = getElement(method_params, "include_ps_dis"),
                       ps_method = getElement(method_params, "ps_method_dis"),
                       oreg_method = getElement(method_params,"oreg_method_dis"),
                       X_names = X_names,
                       offset = getElement(method_params, "offset"))
-
-    ite_dis <- ite_list_dis[["ite"]]
   } else {
     logger::log_info("Using the provided ITE estimations...")
   }
@@ -185,18 +183,15 @@ cre <- function(y, z, X,
   # Estimate ITE
   if (is.null(ite)) {
     logger::log_info("Estimating ITE...")
-    ite_list_inf <- estimate_ite(y = y_inf, z = z_inf, X = X_inf,
+    ite_inf <- estimate_ite(y = y_inf, z = z_inf, X = X_inf,
                       ite_method = getElement(method_params, "ite_method_inf"),
                       include_ps = getElement(method_params, "include_ps_inf"),
                       ps_method = getElement(method_params, "ps_method_inf"),
                       oreg_method = getElement(method_params, "oreg_method_inf"),
                       X_names = X_names,
                       offset = getElement(method_params,"offset"))
-
-    ite_inf <- ite_list_inf[["ite"]]
-    sd_ite_inf <- ite_list_inf[["sd_ite"]]
   } else {
-    sd_ite_inf <- NULL
+    logger::log_info("Using the provided ITE estimations...")
   }
 
   # Generate rules matrix