+++ R/estimate_ite_bart.R
@@ -13,8 +13,7 @@
 #' covariate in ITE estimation.
 #' @param ps_method Method for the estimation of the propensity score.
 #'
-#' @return A list of ITE estimates and standard deviations for the ITE
-#' estimates.
+#' @return A list of ITE estimates.
 #'
 #' @note The number of samples and the number of burn are set by default equal
 #' to 500.
@@ -33,6 +32,5 @@ estimate_ite_bart <- function(y, z, X, include_ps, ps_method) {
 
   pd_ite <- bartCause::extract(bart_fit, type = "ite")
   ite <- apply(pd_ite, 2, mean)
-  sd_ite <- apply(pd_ite, 2, stats::sd)
-  return(list(ite, sd_ite))
+  return(ite)
 }