+++ tests/testthat/test-discover_rules.R
@@ -42,7 +42,7 @@ test_that("discover_rules works as expected!", {
   check_hyper_params(params = hyper_params)
 
   # Estimate ITE
-  ite_list <- estimate_ite(y = y, z = z, X = X,
+  ite <- estimate_ite(y = y, z = z, X = X,
                 ite_method = getElement(method_params, "ite_method_dis"),
                 binary_outcome = getElement(method_params, "binary_outcome"),
                 include_ps = getElement(method_params, "include_ps_dis"),
@@ -50,7 +50,6 @@ test_that("discover_rules works as expected!", {
                 oreg_method = getElement(method_params, "oreg_method_dis"),
                 X_names = X_names,
                 offset = getElement(method_params, "offset"))
-  ite <- ite_list[["ite"]]
 
   # Generate Causal Decision Rules
   select_rules <- discover_rules(X, ite, method_params, hyper_params)