+++ tests/testthat/test-generate_rules.R
@@ -27,12 +27,10 @@ test_that("generate_rules works as expected!", {
   z <- as.matrix(z)
 
   # Step 2: Estimate ITE
-  ite_list <- estimate_ite(y, z, X, ite_method,
+  ite <- estimate_ite(y, z, X, ite_method,
                            include_ps = include_ps,
                            ps_method = ps_method,
-                           or_method = or_method,
-                           random_state = 298)
-  ite <- ite_list[["ite"]]
+                           or_method = or_method)
 
   # Correct outputs
   rules <- generate_rules(X, ite, intervention_vars, ntrees_rf, ntrees_gbm,
@@ -63,11 +61,10 @@ test_that("generate_rules works as expected!", {
   z <- as.matrix(z)
 
   # Step 2: Estimate ITE
-  ite_list <- estimate_ite(y, z, X, ite_method,
+  ite <- estimate_ite(y, z, X, ite_method,
                            include_ps = include_ps,
                            ps_method = ps_method,
                            or_method = or_method)
-  ite <- ite_list[["ite"]]
 
   # Correct outputs
   rules <- generate_rules(X, ite, intervention_vars, ntrees_rf, ntrees_gbm,